+++ src/leiningen/new/re_frame.clj
@@ -11,6 +11,6 @@
            re-frame-assets)
      (-> options
          (assoc :re-frame true)
-         (append-options :dependencies [['re-frame "0.10.5"]])
+         (append-options :dependencies [['re-frame "0.10.6"]])
          (append-options :dev-dependencies [['day8.re-frame/re-frame-10x "0.3.3-react16"]]))]
     state))