+++ src/leiningen/new/reagent.clj
@@ -12,7 +12,7 @@
    ['baking-soda "0.2.0"
     :exclusions
     ['cljsjs/react-bootstrap]]
-   ['cljsjs/react-transition-group "2.3.1-0"]
+   ['cljsjs/react-transition-group "2.4.0-0"]
    ['cljsjs/react-popper "0.10.4-0"]])
 
 (defn reagent-features [[assets options :as state]]